rename files extensions to use typescript
