+++ web-client/package.json
@@ -7,6 +7,11 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.15",
     "@fortawesome/free-solid-svg-icons": "^5.7.2",
     "@fortawesome/react-fontawesome": "^0.1.4",
+    "@types/jest": "^24.0.11",
+    "@types/node": "^11.11.0",
+    "@types/react": "^16.8.7",
+    "@types/react-dom": "^16.8.2",
+    "@types/react-fontawesome": "^1.6.4",
     "axios": "^0.18.0",
     "i": "^0.3.6",
     "npm": "^6.8.0",
@@ -14,11 +19,13 @@
     "react": "^16.8.2",
     "react-axios": "^2.0.3",
     "react-dom": "^16.8.2",
+    "react-placeholder": "^3.0.2",
     "react-redux": "^6.0.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.5",
     "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "typescript": "^3.3.3333"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,6 +43,8 @@
     "not op_mini all"
   ],
   "devDependencies": {
+    "@types/react-redux": "^7.0.3",
+    "@types/react-router-dom": "^4.3.1",
     "redux-devtools-extension": "^2.13.8"
   }
 }
