+++ web-client/src/components/fa-library.js
@@ -1,0 +1,9 @@
+import { library } from '@fortawesome/fontawesome-svg-core';
+import { faBars, faTimes, faChevronRight, faChevronLeft, faCheck } from '@fortawesome/free-solid-svg-icons';
+library.add([
+  faBars,
+  faTimes,
+  faChevronRight,
+  faChevronLeft,
+  faCheck,
+]);
