+++ web-client/src/store/topnav.ts
@@ -1,0 +1,3 @@
+export interface TopNavState {
+  isToggleOn: boolean
+}
